prevent to choose wrong acion
